Added base pagination, table, collection controllers
